Don't mark as !Send
Rustbox was definitely Send (it didn't use thread local storage
or reference the stack so it could be safely sent between threads).
However, it wasn't Sync (it couldn't be shared between threads) because
it had interior mutability. So, to fix this, I changed the mutating
methods to take `&mut self` instead of `&self`. This way, it should be
possible to safely send rustbox instances via channels and share them
with mutexes.